package/freescale-imx/imx-gpu-viv: fix build issue with Wayland backend
Just like Jerome had to do with EGL_API_FB back in the days, in order to
compile with Vivante header it is necessary to have some defines setup
as the default behavior of eglplatform.h is to include Xlib.h.

So this patch sets WL_EGL_PLATFORM when the Wayland backend is used.
This option is actually declared in egl.pc but not all packages use
pkg-config properly.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>